chore: add dependabot to configure package updates
Added a new `.github/dependabot.yml` file to configure Dependabot for
monthly package updates. The configuration includes Docker, GitHub
Actions, and npm package ecosystems, with directories specified for apps
and libs. This will ensure that our dependencies are regularly updated
to the latest versions, improving the overall stability and security of
the project.

Signed-off-by: Luciano Douglas <lucianodouglasm2@gmail.com>